Edit Vector.h Iterator.h main.cpp
