+++ DataStructure/CPP_STL/Debug/CPP_STL.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>C:\Users\Keemin\Desktop\STL\DataStructure\Debug\CPP_STL.exe</FullPath>
+      <FullPath>D:\File\STL\DataStructure\Debug\CPP_STL.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
