+++ DataStructure/CPP_STL/Debug/CPP_STL.log
@@ -1,2 +1,1 @@
-﻿  Vector.cpp
-  CPP_STL.vcxproj -> C:\Users\Keemin\Desktop\STL\DataStructure\Debug\CPP_STL.exe
+﻿  CPP_STL.vcxproj -> D:\File\STL\DataStructure\Debug\CPP_STL.exe
