+++ DataStructure/CPP_STL/Debug/CPP_STL.tlog/CPP_STL.lastbuildstate
@@ -1,2 +1,2 @@
-PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.29.30133:TargetPlatformVersion=10.0.19041.0:
-Debug|Win32|C:\Users\Keemin\Desktop\STL\DataStructure\|
+PlatformToolSet=v142:VCToolArchitecture=Native32Bit:VCToolsVersion=14.29.30133:VCServicingVersionCrtHeaders=14.29.30136:TargetPlatformVersion=10.0.20348.0:
+Debug|Win32|D:\File\STL\DataStructure\|
