+++ DataStructure/CPP_STL/Debug/Iterator.hpp.obj.dt.md.json
@@ -1,0 +1,9 @@
+{
+    "Version": "1.1",
+    "Data": {
+        "Source": "d:\\file\\stl\\datastructure\\cpp_stl\\iterator.hpp",
+        "ProvidedModule": "",
+        "ImportedModules": [],
+        "ImportedHeaderUnits": []
+    }
+}
