+++ DataStructure/CPP_STL/Iterator.h
@@ -1,0 +1,86 @@
+#pragma once
+
+template<typename T>
+class Iterator
+{
+public:
+	Iterator() : _ptr(nullptr)
+	{
+
+	}
+
+	Iterator(T* ptr) : _ptr(ptr)
+	{
+
+	}
+
+	Iterator& operator++()
+	{
+		_ptr++;
+		return *this;
+	}
+
+	Iterator operator++(int)
+	{
+		Iterator temp = *this;
+		_ptr++;
+
+		return temp;
+
+	}
+
+	Iterator& operator--()
+	{
+		_ptr--;
+		return *this;
+	}
+
+	Iterator operator--(int)
+	{
+		Iterator temp = *this;
+		_ptr--;
+
+		return temp;
+
+	}
+
+	Iterator operator+(const int count)
+	{
+		Iterator temp = *this;
+		temp._ptr += count;
+		return temp;
+	}
+
+	Iterator operator-(const int count)
+	{
+		Iterator temp = *this;
+		temp._ptr -= count;
+		return temp;
+	}
+
+	bool operator==(const Iterator& right)
+	{
+		return _ptr == right._ptr;
+	}
+
+
+	bool operator!=(const Iterator& right)
+	{
+		return !(*this == right);
+	}
+
+	T& operator*()
+	{
+		return *_ptr;
+	}
+
+	~Iterator()
+	{
+
+	}
+
+
+public:
+	T* _ptr;
+
+};
